pyproject.toml
[project]
name = "glp-1-landscape"
version = "0.1.0"
description = "Example workflow to build a model of the GLP-1RA clinical trial landscape"
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">3.12,<3.14"
dependencies = [
    "biopython (>=1.86,<2.0)",
    "pydantic[email] (>=2.12.3,<3.0.0)",
    "langchain-openai (>=1.0.2,<2.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "jupyter (>=1.1.1,<2.0.0)",
    "pandas (>=2.3.3,<3.0.0)",
    "langchain-core (>=1.0.3,<2.0.0)",
    "falkordb (>=1.2.0,<2.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "langgraph (>=1.0.2,<2.0.0)",
    "langchain (>=1.0.5,<2.0.0)",
    "langchain-tools (>=0.1.34,<0.2.0)",
    "openai (>=2.7.1,<3.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.0,<10.0.0)",
    "pytest-mock (>=3.15.1,<4.0.0)",
    "pytest-cov (>=7.0.0,<8.0.0)"
]
